+++ lib/componants/story_view.dart
@@ -1,7 +1,7 @@
-import 'package:flutter/material.dart';
-import 'package:provider/provider.dart';
-import 'package:complete_story_view/componants/story_widget.dart';
+import 'package:complete_story_view/controller/story_inherited_notifier.dart';
 import 'package:complete_story_view/controller/story_provider.dart';
+import 'package:flutter/material.dart';
+import 'package:complete_story_view/componants/story_widget.dart';
 
 import '../widgets/linear_timer.dart';
 
@@ -11,69 +11,95 @@ class StoryView extends StatefulWidget {
   final Color? barColor;
   final Color? progressCompletedColor;
   final Color? progressColor;
-  const StoryView({
+  StoryView({
     super.key,
     required this.storyChildren,
+
+    ///height of the top progress bar of the story
     this.progressBarHeight,
     this.barColor,
     this.progressColor,
     this.progressCompletedColor,
-  });
+  }) : assert(storyChildren.isNotEmpty, 'storyChildren list is empty');
 
   @override
   State<StoryView> createState() => _StoryViewState();
 }
 
 class _StoryViewState extends State<StoryView> {
-  PageController pageController = PageController(initialPage: 0);
+  late StoryProvider storyProvider;
+
+  bool isDetailsLoading = true;
+
+  @override
+  void initState() {
+    super.initState();
+    WidgetsBinding.instance.addPostFrameCallback((_) {
+      storyProvider = StoryScope.of(context);
+      storyProvider.innerPageController = PageController(
+        initialPage: storyProvider.currentIndex,
+      );
+      isDetailsLoading = false;
+      setState(() {});
+    });
+  }
+
+  @override
+  void didChangeDependencies() {
+    super.didChangeDependencies();
+  }
+
+  @override
+  void dispose() {
+    storyProvider.innerPageController.dispose();
+    super.dispose();
+  }
 
   @override
   Widget build(BuildContext context) {
-    return Consumer<StoryProvider>(
-      builder: (context, state, child) {
-        return GestureDetector(
-          onVerticalDragEnd: (value) {
-            if (value.velocity.pixelsPerSecond.distance > 0 &&
-                value.velocity.pixelsPerSecond.dy > 0) {
-              Navigator.pop(context);
-            }
-          },
-          child: Stack(
-            children: [
-              // if (!state.isDetailsLoading)
-              PageView.builder(
-                physics: NeverScrollableScrollPhysics(),
-                controller: state.innerPageController,
-                itemCount: widget.storyChildren.length,
-                onPageChanged: (value) {
-                  state.updateCurrenIndex(value);
-                  // if (widget.storyChildren.length == value + 1) {
-                  //   log("LAST STORY");
-                  // }
-                },
-                itemBuilder: (BuildContext context, int index) {
-                  return widget.storyChildren[index];
-                },
-              ),
-              if (!state.isDetailsLoading)
+    return isDetailsLoading
+        ? SizedBox.shrink()
+        : GestureDetector(
+            onVerticalDragEnd: (value) {
+              if (value.velocity.pixelsPerSecond.distance > 0 &&
+                  value.velocity.pixelsPerSecond.dy > 0) {
+                Navigator.pop(context);
+              }
+            },
+            child: Stack(
+              children: [
+                // if (!state.isDetailsLoading)
+                PageView.builder(
+                  physics: NeverScrollableScrollPhysics(),
+                  controller: storyProvider.innerPageController,
+                  itemCount: widget.storyChildren.length,
+                  onPageChanged: (value) {
+                    // storyProvider.updateCurrenIndex(value);
+                  },
+                  itemBuilder: (BuildContext context, int index) {
+                    return widget.storyChildren[index];
+                  },
+                ),
                 Padding(
                   padding: const EdgeInsets.only(top: 6, left: 6),
                   child: Row(
                     children: widget.storyChildren.map((e) {
-                      return state.currentIndex ==
+                      return storyProvider.currentIndex ==
                               widget.storyChildren.indexOf(e)
                           ? Expanded(
                               child: LinearTimer(
                                 progressBarHeight: widget.progressBarHeight,
                                 onTimerFinish: () {
-                                  state.onNextStory(
+                                  storyProvider.onNextStory(
                                     widget.storyChildren.length,
                                   );
                                 },
-                                durationMiliseconds: state.durationCurrentStory,
+                                durationMiliseconds:
+                                    storyProvider.durationCurrentStory,
                               ),
                             )
-                          : state.currentIndex > widget.storyChildren.indexOf(e)
+                          : storyProvider.currentIndex >
+                                widget.storyChildren.indexOf(e)
                           ? Expanded(
                               child: Padding(
                                 padding: const EdgeInsets.only(right: 6),
@@ -103,13 +129,14 @@ class _StoryViewState extends State<StoryView> {
                     }).toList(),
                   ),
                 ),
-              if (!state.isDetailsLoading)
                 Row(
                   children: [
                     Expanded(
                       child: GestureDetector(
                         onTap: () {
-                          state.onPrevStory(widget.storyChildren.length);
+                          storyProvider.onPrevStory(
+                            widget.storyChildren.length,
+                          );
                         },
                         child: Container(
                           height: double.infinity,
@@ -120,7 +147,9 @@ class _StoryViewState extends State<StoryView> {
                     Expanded(
                       child: GestureDetector(
                         onTap: () {
-                          state.onNextStory(widget.storyChildren.length);
+                          storyProvider.onNextStory(
+                            widget.storyChildren.length,
+                          );
                         },
                         child: Container(
                           height: double.infinity,
@@ -130,18 +159,8 @@ class _StoryViewState extends State<StoryView> {
                     ),
                   ],
                 ),
-              if (state.isDetailsLoading && state.pageNavi)
-                SizedBox(
-                  height: double.infinity,
-                  width: double.infinity,
-                  child: Center(child: CircularProgressIndicator()),
-                ),
-              if (state.isDetailsLoading)
-                Center(child: CircularProgressIndicator(color: Colors.white54)),
-            ],
-          ),
-        );
-      },
-    );
+              ],
+            ),
+          );
   }
 }