+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 0.0.4
+
+* fix: removed provider dependancy
+* features : [imageStoryDuration] for set duration of an image story
+* features : [onComplete] triggers when last story gets completed
+
 ## 0.0.3
 
 * fix: removed image dependancy